+++ Cargo.toml
@@ -123,6 +123,7 @@ tokio-core = { version = "0.1", optional = true }
 tokio = { version = "0.2.22", features = ["io-util", "rt-threaded", "stream", "tcp"] }
 unwrap = "1.2"
 uuid = { version = "0.7", features = ["serde", "v4"] }
+wasm-timer = "0.2.4"
 winapi = "0.3"
 
 wasm-bindgen = { version = "0.2.50", features = ["serde-serialize", "nightly"], optional = true }
