+++ mm2src/docker_tests.rs
@@ -2,6 +2,7 @@
 #![feature(test)]
 #![test_runner(docker_tests_runner)]
 #![feature(drain_filter)]
+#![feature(hash_raw_entry)]
 #![feature(non_ascii_idents)]
 #![recursion_limit = "512"]
