+++ mm2src/mm2_bin.rs
@@ -2,6 +2,7 @@
 #![feature(drain_filter)]
 #![recursion_limit = "512"]
 #![feature(test)]
+#![feature(hash_raw_entry)]
 
 #[macro_use] extern crate common;
 #[macro_use] extern crate fomat_macros;
